move weight quant file
